Update parameters parsing
